+++ CHANGELOG.md
@@ -240,7 +240,7 @@ The following changes have been implemented but not released yet:
 
 ### Bugfixes
 
-- `getWellKnownSolid` fetches well known solid from the server's root by default.
+- `getWellKnownSolid` fetches well-known solid from the server's root by default.
 
 ## [1.22.0] - 2022-04-26